Fix : userId, keywordId 타입 수정

+++ src/main/java/com/know_who_how/main_server/auth/AuthRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AuthRepository extends JpaRepository<User, Integer> {
+public interface AuthRepository extends JpaRepository<User, Long> {
 
 
     Optional<User> findByLoginId(String loginId);
