+++ src/main/java/com/know_who_how/main_server/global/entity/Keyword/UserKeywordId.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 @NoArgsConstructor
 public class UserKeywordId implements Serializable {
 
-    private Integer user;
+    private Long user;
     private Long keyword;
 
     @Override
